Changes to build flags, primarily. Alterations to default configuration settings. Making the VM far more lax regarding permissions.

+++ src/java.desktop/windows/native/libawt/windows/awt_Menu.h
@@ -74,7 +74,7 @@ public:
 
     /*for multifont menu */
     BOOL IsTopMenu();
-    virtual AwtMenuItem* GetItem(jobject target, long index);
+    virtual AwtMenuItem* GetItem(jobject target, jint index);
 
     virtual int CountItem(jobject target);
